Added prefabrication for configuration create action
Necessary to provide a prefabricated action which can be used by upstream projects for creating a project's configuration.

Use of `ARGV.clear` is necessary since `Kernel#gets` reads from it. So if anything, *dirty*, is in `ARGV` then it'll mess up how `Kernel#gets` processes user input. Even though this is a destructive mutation, it's OK in this situation since excecution of this action is the last thing that happens before the CLI exits. In the future, prompting for user input might need to be purely a CLI flag to avoid prompting altogether.

💡 This action pairs well with the following gems:

- [Etcher](https://alchemists.io/projects/etcher)
- [Runcom](https://alchemists.io/projects/runcom)