Fixed reusable commands without handles
Necessary to resolve an inconsistency with reusable commands which didn't require a handle like actions and inline commands do. This change ensures reusable commands must always have a handle which also fixes an issue with node presenters. This also ensures reusable commands accept only strings, like handles, which is also consistent with how inline commands work.

RuboCop was updated to ignore the long example because breaking up the rendering of the help documentation would hurt spec readability in this case so keeping all of this together is preferred.